Clean up whitespace and trailing parens
It's quite typical to find that you have a good few parens at the end of
a block. That's fine. In most cases your editor (I switch between Vim
and Emacs 20/80%) should be able to help you here.

As lisp code is just a bunch of list literals you just have to make sure
the body of your blocks/expressions are readable. The trailing parens
shouldn't matter.