(max-value ..) should be thought of as a list op.
If you think of your function as different operations on a list, rather
than attempting to find the solution in place you can create something
far more readable.

In this case, I split out the actions as: sort, by a comparator (>) and
then pop the top value off.

This ends up being very similar for the opposite case.